see changelog
git-svn-id: svn://svn.hafro.is/jminor/trunk@6650 51816742-9f83-410d-85ed-e7bcf7042953

+++ resources/demos/chinook/scripts/ddl.sql
@@ -91,7 +91,7 @@ CREATE TABLE Chinook.Invoice
     BillingState VARCHAR(40) ,
     BillingCountry VARCHAR(40) ,
     BillingPostalCode VARCHAR(10) ,
-    Total DOUBLE, --NOT NULL,
+    Total DECIMAL(12, 2), --NOT NULL,
     CONSTRAINT PK_Invoice PRIMARY KEY (InvoiceId)
 );
